pyproject.toml
[project]
name = "cuckoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "dynaconf>=3.2.12",
    "elasticsearch>=8.17.0,<9.0.0",
    "fastapi>=0.128.8",
    "httpx>=0.27.0",
    "injector>=0.24.0",
    "pyvespa>=0.49.0",
    "redis>=7.1.1",
    "rocketry>=2.5.1",
    "sentence-transformers>=3.3.1",
    "torch>=2.5.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
cuckoo = "src.main:main"
deploy-vespa = "scripts.deploy_vespa:deploy"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 100
# Enable sorting (I) and removal of unused imports/variables (F)
lint.select = ["E", "F", "I"]
lint.fixable = ["ALL"]

[tool.ruff.lint.isort]
# Like Go's grouping: keeps your own packages separate from standard libs
known-first-party = ["src"]
